Refactor Empty Epsilon integration structure
Refactor into TypeScript and split the functionality into separate files
- Express API routes
- Client
- HTTP Server emulator
- State synchronization

Also adds initial support for synchronizing landing pad states.
